update Editor recent list view: only delete file on disk when it is absolute path

+++ app/src/main/java/com/catpuppyapp/puppygit/screen/content/homescreen/innerpage/EditorInnerPage.kt
@@ -1298,10 +1298,12 @@ fun EditorInnerPage(
                             FileOpenHistoryMan.remove(it.file.path.ioPath)
 
                             if(deleteFileOnDisk) {
-                                try {
-                                    File(it.file.path.ioPath).delete()
-                                }catch (e: Exception) {
-                                    MyLog.w(TAG, "remove file failed: ioPath=${it.file.path.ioPath}, err=${e.localizedMessage}")
+                                if(it.file.path.ioPathType == PathType.ABSOLUTE) {
+                                    try {
+                                        File(it.file.path.ioPath).delete()
+                                    }catch (e: Exception) {
+                                        MyLog.w(TAG, "remove file failed: ioPath=${it.file.path.ioPath}, err=${e.localizedMessage}")
+                                    }
                                 }
                             }
                         }
